update Note-related DTOs to use Long for ID fields

+++ core/network/src/commonMain/kotlin/com/mifos/core/network/dto/note/CreateNoteResponseDto.kt
@@ -13,7 +13,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class CreateNoteResponseDto(
-    val officeId: Int? = null,
-    val clientId: Int? = null,
-    val resourceId: Int? = null,
+    val officeId: Long? = null,
+    val clientId: Long? = null,
+    val resourceId: Long? = null,
 )
